General code fixing

+++ src/Controllers/Collection/Delete/index.js
@@ -4,11 +4,12 @@ import asyncHandler from 'express-async-handler';
 
 export default asyncHandler(async (req, res) => {
     const { user } = req;
-    const { Name = new CustomError('Collection "Name" is required'), force = false } = req.body;
-    const _collection = new BaseCollection({ Name, UserId: user.id });
+    const { CollectionName, force = false } = req.body;
+    if (!CollectionName) throw new CustomError('Collection "Name" is required');
+    const _collection = new BaseCollection({ Name: CollectionName, UserId: user.id });
     await _collection.Delete(force);
     res.json({
         success: true,
-        message: `Collection ${Name} deleted successfully`
+        message: `Collection ${CollectionName} deleted successfully`
     });
 });
